fix: Novo arquivo com classes auxiliadores para o webhooks
